Add optional additional entities to authz request
Added an option to pass in entities to authorization requests that would
not overwrite the configuration already stored in memory, but modify it
by adding new entities to it.

Should be used if your use case has a large amount of entities that is
infeasible to store in memory and you have easy access to all relevant
entities when making the authorization request.